CODIGO/requirements.txt
pillow 
rembg